pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentic-ai"
version = "0.1.0"
description = "Modular Agentic AI System with LangChain v1.0 and LangGraph"
authors = [
    {name = "Agentic AI Team", email = "team@agentic-ai.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core LangChain dependencies
    "langchain>=1.0.1",
    "langchain-core>=0.2.0",
    "langchain-community>=0.4",
    "langgraph>=1.0.1",
    # LLM Provider integrations
    "langchain-ollama>=1.0.0",
    "langchain-google-genai>=3.0.0",
    "langchain-openai>=1.0.0",
    # MCP and tools
    "langchain-mcp-adapters>=0.1.0",
    "mcp>=1.0.0",
    # Core utilities
    "pydantic>=2.12.3",
    "pyyaml>=6.0",
    "requests>=2.32.5",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    # Ollama
    "ollama>=0.6.0",
    # Web framework (기존 프로젝트용)
    "fastapi>=0.119.1",
    "uvicorn>=0.38.0",
    "deepagents>=0.1.4",
    "pandas>=2.3.3",
    "langfuse>=3.9.0",
    "langgraph-cli[inmem]>=0.4.7",
    "langchain-tavily>=0.2.13",
    "langgraph-api>=0.5.2",
    "pytest>=8.4.2",
    "docker>=7.1.0",
    "langchain-experimental>=0.4.0",
    "ipython>=9.7.0",
    "psutil>=7.1.3",
    "unstructured>=0.18.21",
    "pymupdf>=1.26.6",
    "pypdf>=6.4.0",
    "pdfminer-six>=20221105",
    "pi-heif>=1.1.1",
    "unstructured-inference>=1.1.2",
    "pdf2image>=1.17.0",
    "unstructured-pytesseract>=0.3.15",
    "pytesseract>=0.3.13",
    "docling>=2.65.0",
    "mineru[core,lmdeploy,vllm]>=2.6.5",
    "transformers==4.56.1",
    "gdown>=5.2.0",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0"
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.uv.workspace]
members = [
    "docling_test",
]
